mst: cleanup files
